pyproject.toml
[project]
name = "banalyze-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.3.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
